server/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.16.0
